+++ CE4703_Assignment_2/Game.h
@@ -30,14 +30,14 @@
  * @param numberOfPlayers How many players participating in the game
  * @param numberPacks Number of 52-card packs to use
  */
-void startGame(int numberPlayers, int numberPacks);
+void startGame(int numberPlayers, long long numberPacks);
 
 /**
  * @brief Fill a deck with standard 52-card pack(s)
  * @param d Pointer to the deck to fill
  * @param packs Number of 52-card packs to add
  */
-void fillDeck(Deck* d, int packs);
+void fillDeck(Deck* d, long long packs);
 
 /**
  * @brief Reshuffle played cards back into hidden deck, keeping top card
